+++ lib/src/a1_range.dart
@@ -37,45 +37,32 @@ class A1Range implements Comparable<A1Range> {
   final Object? tag;
 
   /// Private contructor
-  A1Range._(A1Partial from, A1Partial to, {this.anchor, this.tag}) {
-    // normalise the reverse diagonal to always have
-    // eg. A2:B3
-    switch ((from.a1, to.a1)) {
-      case (
-            A1(column: var columnFrom, row: var rowFrom),
-            A1(column: var columnTo, row: var rowTo)
-          )
-          when columnTo < columnFrom:
-        this.from = A1Partial.fromVector(columnTo, rowFrom);
-        this.to = A1Partial.fromVector(columnFrom, rowTo);
-      default:
-        this.from = from;
-        this.to = to;
-    }
-  }
+  A1Range._(this.from, this.to, {this.anchor, this.tag});
 
   /// Creates a range from two supplied A1Partials for from and to
   /// ensures 'from' is less than or equal to 'to' when not all
   /// if the from is an A1 is will be treated as the anchor
   factory A1Range.fromPartials(A1Partial from, A1Partial to, {Object? tag}) {
-    if ((from.isAll || to.isAll) || from <= to) {
-      return A1Range._(from, to, anchor: from.a1, tag: tag);
-    } else {
-      return A1Range._(to, from, anchor: from.a1, tag: tag);
-    }
+    final (left, top, right, bottom) = _rectify(from, to, min: -1);
+    final fromNormal = A1Partial.fromVector(
+        _minPartial(left, right), _minPartial(top, bottom));
+    final toNormal = A1Partial.fromVector(
+        _maxPartial(left, right), _maxPartial(top, bottom));
+
+    return A1Range._(fromNormal, toNormal, anchor: from.a1, tag: tag);
   }
 
   /// Creates a range from two supplied A1s for from and to
   /// ensures 'from' is less than or equal to 'to'
   /// the fromA1 will be used as the anchor in the range
   factory A1Range.fromA1s(A1 fromA1, A1 toA1, {Object? tag}) {
-    final from = A1Partial.fromA1(fromA1);
-    final to = A1Partial.fromA1(toA1);
-    if (from <= to) {
-      return A1Range._(from, to, anchor: fromA1, tag: tag);
-    } else {
-      return A1Range._(to, from, anchor: fromA1, tag: tag);
-    }
+    final (left, top, right, bottom) =
+        _rectify(A1Partial.fromA1(fromA1), A1Partial.fromA1(toA1), min: -1);
+    final fromNormal = A1Partial.fromVector(
+        _minPartial(left, right), _minPartial(top, bottom));
+    final toNormal = A1Partial.fromVector(
+        _maxPartial(left, right), _maxPartial(top, bottom));
+    return A1Range._(fromNormal, toNormal, anchor: fromA1, tag: tag);
   }
 
   /// Parses a string containing an A1Range literal into an A1Range.
@@ -157,9 +144,18 @@ class A1Range implements Comparable<A1Range> {
     A1? anchor,
     Object? tag,
   }) {
-    return A1Range._(
+    final (left, top, right, bottom) = _rectify(
       from ?? this.from,
       to ?? this.to,
+      min: -1,
+    );
+    final fromNormal = A1Partial.fromVector(
+        _minPartial(left, right), _minPartial(top, bottom));
+    final toNormal = A1Partial.fromVector(
+        _maxPartial(left, right), _maxPartial(top, bottom));
+    return A1Range._(
+      fromNormal,
+      toNormal,
       anchor: anchor ?? this.anchor,
       tag: tag ?? this.tag,
     );
@@ -249,86 +245,16 @@ class A1Range implements Comparable<A1Range> {
   }
 
   /// is the row in this range
-  bool hasRow(int row) => switch ((from.a1, to.a1)) {
-        // A1:A10
-        (A1(row: var fromRow), A1(row: var toRow))
-            when row >= fromRow && row <= toRow =>
-          true,
-
-        // :A1
-        (_, A1(row: var toRow)) when from.isAll && row <= toRow => true,
-
-        // A:B2
-        (null, A1(row: var toRow))
-            when from.row != null && row >= from.row! && row <= toRow =>
-          true,
-
-        // A1:
-        (A1(row: var fromRow), _) when row >= fromRow && to.isAll => true,
-
-        // A1:1
-        (A1(row: var fromRow), null)
-            when row >= fromRow && to.row != null && row <= to.row! =>
-          true,
-
-        // A:B
-        (null, null) when from.row == null && to.row == null => true,
-
-        // 1:1
-        (null, null)
-            when from.row != null &&
-                row >= from.row! &&
-                to.row != null &&
-                row <= to.row! =>
-          true,
-        (_, _) when from.isAll && to.isAll => true,
-        _ => false,
-      };
+  bool hasRow(int row) {
+    final (fromR, toR) = (from.row ?? 0, to.row ?? _maxInt);
+    return row >= fromR && row <= toR;
+  }
 
   /// is the column in this range
-  bool hasColumn(int column) => switch ((from.a1, to.a1)) {
-        // A1:A10
-        (A1(column: var fromColumn), A1(column: var toColumn))
-            when column >= fromColumn && column <= toColumn =>
-          true,
-
-        // :B2
-        (_, A1(column: var toColumn)) when from.isAll && column <= toColumn =>
-          true,
-
-        // 2:B2
-        (null, A1(column: var toColumn))
-            when from.column != null &&
-                column >= from.column! &&
-                column <= toColumn =>
-          true,
-
-        // A1:
-        (A1(column: var fromColumn), _) when column >= fromColumn && to.isAll =>
-          true,
-
-        // A1:A
-        (A1(column: var fromColumn), null)
-            when column >= fromColumn &&
-                to.column != null &&
-                column <= to.column! =>
-          true,
-
-        // 1:2
-        (null, null) when from.column == null && to.column == null => true,
-
-        // A:A
-        (null, null)
-            when from.column != null &&
-                column >= from.column! &&
-                to.column != null &&
-                column <= to.column! =>
-          true,
-
-        // <empty>
-        (_, _) when from.isAll && to.isAll => true,
-        _ => false,
-      };
+  bool hasColumn(int column) {
+    final (fromC, toC) = (from.column ?? 0, to.column ?? _maxInt);
+    return column >= fromC && column <= toC;
+  }
 
   /// Select the left border strip for this range
   A1Range get leftBorder =>
@@ -433,6 +359,38 @@ class A1Range implements Comparable<A1Range> {
     return result;
   }
 
+  /// the number of columns between the start and end of this range
+  /// uses the maximum integer of the system for unbound columns starting
+  /// at 0, this will always be greater than or equal to 1
+  int get columnSpan => (right - left).a1PIncrement ?? _maxInt;
+
+  /// the number of rows between the start and end of this range
+  /// uses the maximum integer of the system for unbound rows starting
+  /// at 0, this will always be greater than or equal to 1
+  int get rowSpan => (bottom - top).a1PIncrement ?? _maxInt;
+
+  /// the from.column or 0
+  int get left => (from.column ?? 0);
+
+  /// the from.row or 0
+  int get top => (from.column ?? 0);
+
+  /// the to.column or maximum integer for the system
+  int get right => (to.column ?? _maxInt);
+
+  /// the to.row or maximum integer for the system
+  int get bottom => (to.row ?? _maxInt);
+
+  // utility to move [A1Range] in to rectable-like scalars
+  static (int, int, int, int) _rectify(A1Partial from, A1Partial to,
+          {int min = 0}) =>
+      (
+        from.column ?? min,
+        from.row ?? min,
+        to.column ?? _maxInt,
+        to.row ?? _maxInt,
+      );
+
   /// Return the [A1Range]s that is left after subtracting
   /// the other [A1Range]. Maintains the tag
   ///
@@ -442,18 +400,9 @@ class A1Range implements Comparable<A1Range> {
     }
 
     List<A1Range> result = [];
-    final (int x1, int y1, int x2, int y2) = (
-      from.column ?? 0,
-      from.row ?? 0,
-      to.column ?? _maxInt,
-      to.row ?? _maxInt,
-    );
-    final (int otherX1, int otherY1, int otherX2, int otherY2) = (
-      other.from.column ?? 0,
-      other.from.row ?? 0,
-      other.to.column ?? _maxInt,
-      other.to.row ?? _maxInt,
-    );
+    final (int x1, int y1, int x2, int y2) = _rectify(from, to);
+    final (int otherX1, int otherY1, int otherX2, int otherY2) =
+        _rectify(other.from, other.to);
     if (x1 < otherX1) {
       result.add(A1Range.fromPartials(
         A1Partial.fromVector(x1.a1P, y1.a1P),
@@ -495,8 +444,8 @@ class A1Range implements Comparable<A1Range> {
     }
   }
 
-  int? _maxPartial(int one, int two) => one >= two ? one.a1P : two.a1P;
-  int? _minPartial(int one, int two) => one <= two ? one.a1P : two.a1P;
+  static int? _maxPartial(int one, int two) => one >= two ? one.a1P : two.a1P;
+  static int? _minPartial(int one, int two) => one <= two ? one.a1P : two.a1P;
 
   /// Return the [A1Range] or null where this and other intersect
   /// updates the tag as [this.tag, other.tag]
@@ -521,18 +470,10 @@ class A1Range implements Comparable<A1Range> {
     } else if (other == all) {
       return copyWith(tag: intersectTag);
     }
-    final (int left, int top, int right, int bottom) = (
-      from.column ?? -1,
-      from.row ?? -1,
-      to.column ?? _maxInt,
-      to.row ?? _maxInt,
-    );
-    final (int otherLeft, int otherTop, int otherRight, int otherBottom) = (
-      other.from.column ?? -1,
-      other.from.row ?? -1,
-      other.to.column ?? _maxInt,
-      other.to.row ?? _maxInt,
-    );
+    final (int left, int top, int right, int bottom) =
+        _rectify(from, to, min: -1);
+    final (int otherLeft, int otherTop, int otherRight, int otherBottom) =
+        _rectify(other.from, other.to, min: -1);
 
     return A1Range.fromPartials(
       A1Partial.fromVector(
@@ -549,18 +490,9 @@ class A1Range implements Comparable<A1Range> {
     if (this == all || other == all) {
       return true;
     }
-    final (int left, int top, int right, int bottom) = (
-      from.column ?? 0,
-      from.row ?? 0,
-      to.column ?? _maxInt,
-      to.row ?? _maxInt,
-    );
-    final (int otherLeft, int otherTop, int otherRight, int otherBottom) = (
-      other.from.column ?? 0,
-      other.from.row ?? 0,
-      other.to.column ?? _maxInt,
-      other.to.row ?? _maxInt,
-    );
+    final (int left, int top, int right, int bottom) = _rectify(from, to);
+    final (int otherLeft, int otherTop, int otherRight, int otherBottom) =
+        _rectify(other.from, other.to);
     if (right < otherLeft || otherRight < left) {
       return false;
     }